+++ test/raw.js
@@ -284,7 +284,7 @@ describe('bodyParser.raw()', function(){
       test.set('Content-Encoding', 'nulls')
       test.set('Content-Type', 'application/octet-stream')
       test.write(new Buffer('000000000000', 'hex'))
-      test.expect(415, 'unsupported content encoding', done)
+      test.expect(415, 'unsupported content encoding "nulls"', done)
     })
   })
 })